requirements.txt
aiohttp
feedparser
nest_asyncio
python-telegram-bot==20.3
